replace swayidle with hypridle
Swayidle would trigger when videos are running, or when playing a game
with a controller.
Hypridle at least doesn't seem to have the same problem when videos are
running. I will have to see later if the same applies to the controller.

Only includes a conf.example file in the repo, since some of the things
specified would be individual pr. computer. (device names used in
commands as an example)